自定义餐品食材重量保存重量历史

+++ src/main/java/com/nutrition/nutritionservice/dao/UserHistoricalCuisineDao.java
@@ -13,7 +13,7 @@ public interface UserHistoricalCuisineDao {
 
     UserHistoricalCuisineVo selectLastCuisineByUuid(String uuid);
 
-    int insert(UserHistoricalCuisineVo record);
+    long insert(UserHistoricalCuisineVo record);
 
     List<UserHistoricalCuisineVo> selectByUuid(String uuid);
